security-automation-co-pilot/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
msal==1.26.0
azure-identity==1.15.0
azure-mgmt-security==5.0.0
azure-mgmt-sentinel==4.0.0
azure-storage-blob==12.19.0
azure-cosmos==4.5.1
pandas==2.2.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
requests==2.31.0
aiohttp==3.9.3
motor==3.3.2
beanie==1.25.0

# Azure Functions and Automation
azure-functions==1.18.0
azure-mgmt-logic==10.0.0
azure-mgmt-network==25.1.0
azure-mgmt-compute==30.4.0
azure-mgmt-resource==23.0.1

# Notification and Templates
jinja2==3.1.2
msgraph-sdk==1.0.0

# Monitoring
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1 